refactor: 클래스별 용도에 맞게 package 생성하였음
