+++ src/main/java/racingcar/model/RacingCar.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
