+++ src/test/java/racingcar/RacingCarsTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import racingcar.model.RacingCars;
+
 class RacingCarsTest {
 	RacingCars racingCars;
